+++ src/commands/hardis/project/deploy/sources/dx.ts
@@ -1,18 +1,16 @@
 /* jscpd:ignore-start */
 
 import { flags, SfdxCommand } from '@salesforce/command';
-import { Messages, SfdxError } from '@salesforce/core';
+import { Messages} from '@salesforce/core';
 import { AnyJson } from '@salesforce/ts-types';
-import * as c from 'chalk';
 import * as fs from 'fs-extra';
 import * as os from 'os';
 import * as path from 'path';
 import * as sortArray from 'sort-array';
 import { MetadataUtils } from '../../../../../common/metadata-utils';
-import { execCommand, uxLog } from '../../../../../common/utils';
-import { analyzeDeployErrorLogs } from '../../../../../common/utils/deployTips';
+import {  uxLog } from '../../../../../common/utils';
 import { getConfig } from '../../../../../config';
-import * as xml2js from 'xml2js';
+import { forceSourceDeploy } from '../../../../../common/utils/deployUtils';
 
 // Initialize Messages with the current plugin directory
 Messages.importMessagesDirectory(__dirname);
@@ -40,6 +38,10 @@ export default class DxSources extends SfdxCommand {
       options: ['NoTestRun', 'RunSpecifiedTests', 'RunLocalTests', 'RunAllTestsInOrg'],
       description: messages.getMessage('testLevel')
     }),
+    packagexml: flags.string({
+      char: 'p',
+      description:"Path to package.xml containing what you want to deploy in target org"
+    }),
     debug: flags.boolean({
       char: 'd',
       default: false,
@@ -62,6 +64,7 @@ export default class DxSources extends SfdxCommand {
     this.configInfo = await getConfig('branch');
     const check = this.flags.check || false;
     const testlevel = this.flags.testlevel || 'RunLocalTests';
+    const packageXml = this.flags.packagexml || null ;
     this.debugMode = this.flags.debug || false;
 
     // Install packages
@@ -84,102 +87,14 @@ export default class DxSources extends SfdxCommand {
     }
 
     // Deploy sources
-    const packageXmlFile =
+    const packageXmlFile = packageXml ||
       process.env.PACKAGE_XML_TO_DEPLOY ||
         this.configInfo.packageXmlToDeploy ||
         (fs.existsSync('./manifest/package.xml')) ? './manifest/package.xml' :
         './config/package.xml';
-    const splitDeploymentPackageXmls = await this.buildDeploymentPackageXmls(packageXmlFile,check);
-    const messages = [];
-    for (const packageXml of splitDeploymentPackageXmls) {
-      const packageXmlFile = packageXml.packageXmlFile ;
-      uxLog(this,c.cyan(`Deploying ${c.bold(packageXml.label)} package: ${packageXmlFile} ...`));
-      if (packageXmlFile.waitBefore) {
-        await new Promise(resolve => setTimeout(resolve, packageXmlFile.waitBefore * 1000));
-      }
-      const deployCommand = `sfdx force:source:deploy -x ${packageXmlFile}` +
-        ' --wait 60' +
-        ' --ignorewarnings' + // So it does not fail in for objectTranslations stuff
-        ` --testlevel ${testlevel}` +
-        (check ? ' --checkonly' : '') +
-        (this.debugMode ? ' --verbose' : '');
-        let deployRes ;
-      try {
-        deployRes = await execCommand(deployCommand, this, { output: true, debug: this.debugMode, fail: true })
-      } catch (e) {
-        const {tips} = analyzeDeployErrorLogs(e.error.stdout + e.error.stderr);
-        uxLog(this,c.red("Sadly there has been Deployment error(s)"));
-        uxLog(this,c.yellow(tips.map((tip:any) => tip.tip).join("\n")));
-        uxLog(this,c.yellow(`You may${tips.length > 0?' also':''} copy-paste errors on google to find how to solve the deployment issues :)`));
-        throw new SfdxError('Deployment failure. Check messages above');
-      }
-      let message = '';
-      if (deployRes.status === 0) {
-        message = `[sfdx-hardis] Successfully ${check ? 'checked deployment of' : 'deployed'} sfdx project sources to Salesforce org`;
-        uxLog(this, c.green(message));
-      } else {
-        message = '[sfdx-hardis] Unable to deploy sfdx project sources to Salesforce org';
-        uxLog(this, c.red(deployRes.errorMessage));
-      }
-      if (packageXmlFile.waitAfter) {
-        await new Promise(resolve => setTimeout(resolve, packageXmlFile.waitAfter * 1000));
-      }
-      messages.push(message);
-    }
+    const {messages} = await forceSourceDeploy(packageXmlFile,check,testlevel);
+
     return { orgId: this.org.getOrgId(), outputString: messages.join("\n") };
-  }
-
-  // In some case we can not deploy the whole package.xml, so let's split it before :)
-  private async buildDeploymentPackageXmls(packageXmlFile: string,check: boolean): Promise<any[]> {
-    const packageXmlString = await fs.readFile(packageXmlFile);
-    const packageXml = await xml2js.parseStringPromise(packageXmlString);
-    // Check for empty package.xml
-    if (!(packageXml && packageXml.Package && packageXml.Package.types && packageXml.Package.types.length > 0)) {
-      uxLog(this,'Empty package.xml: nothing to deploy')
-      return [];
-    }
-    const deploymentPlanFile = path.join(path.dirname(packageXmlFile),'deploymentPlan.json');
-    // Build list of package.xml according to plan
-    if (fs.existsSync(deploymentPlanFile) && !check) {
-      // Read deployment plan
-      const deploymentPlanFileXmlString = await fs.readFile(deploymentPlanFile);
-      const deploymentPlan = await xml2js.parseStringPromise(deploymentPlanFileXmlString);
-      // Copy main package.xml
-      const tmpDeployDir = path.join(os.tmpdir(),'sfdx-hardis-deploy');
-      await fs.ensureDir(tmpDeployDir);
-      const mainPackageXmlCopyFileName = path.join(tmpDeployDir,'mainPackageXml');
-      await fs.copy(packageXmlFile,mainPackageXmlCopyFileName);
-      const mainPackageXmlItem = {
-        label: 'main',
-        packageXmlFile: mainPackageXmlCopyFileName,
-        order: 0
-      }
-      const packageXmlItems = [mainPackageXmlItem];
-      // Remove other package.xml items from main package.xml
-      for (const separatePackageXml of deploymentPlan.packages) {
-        uxLog(this,c.cyan(`Removing ${separatePackageXml.packageXmlFile} content from main package.xml`));
-        const removePackageXmlCommand = 'sfdx essentials:packagexml:remove' +
-        ` --packagexml ${mainPackageXmlCopyFileName}` +
-        ` --removepackagexml ${separatePackageXml.packageXmlFile}` +
-        ` --outputfile ${mainPackageXmlCopyFileName}`;
-        await execCommand(removePackageXmlCommand, this, { fail: true, debug: this.debugMode });
-        packageXmlItems.push(separatePackageXml);
-      }
-
-      // Sort in requested order
-      const packageXmlItemsSorted = sortArray(packageXmlItems, {
-        by: ['order','label'],
-        order: ['asc','asc']
-      });
-      return packageXmlItemsSorted ;
-    }
-    // No transformation: return initial package.xml file
-    return [
-      {
-        label: 'main',
-        packageXmlFile: packageXmlFile
-      }
-    ]
   }
 
 }